refactor(targets): split model families out of a shared glu_transformer target (review)
Per PR #987 review: no switch-behavior inside llama8b, no cartesian-product
configs, no inline model-family dispatch in evals.

- targets/glu_transformer.py (+ _sharding): the SHARED HF GLU-transformer
  machinery — site grammar, FrozenAttn/GLULayer/GLUDecomposedModel (renamed
  from Llama*), scan/masked-forward engine, HF loading parameterized by the
  family's attn loader + inv_freq. FrozenAttn gains a `_prep_qk` pre-RoPE
  hook (identity) and a target-owned `pattern` recipe; GLUDecomposedModel
  gains `attn_pattern(q_site, q_flat, k_flat)`.
- targets/llama8b.py: the Llama family file — llama31_8b_config (vendored
  LlamaConfig, llama3 rope) + load_decomposed_llama_from_hf.
- targets/qwen3_8b.py: the Qwen3 family file — Qwen3FrozenAttn with REQUIRED
  q_norm/k_norm/eps fields overriding _prep_qk (QK-norm keyed by attn class,
  never a flag; invalid states unrepresentable), qwen3_8b_config (GLUConfig),
  load_decomposed_qwen3_from_hf. vendored_jax/llama.py reverted untouched.
- attn_patterns_eval inverted: the recipe is target-owned behind a
  runtime_checkable AttnPatternModel protocol; the eval no longer imports any
  concrete target and the match-on-class dispatch is gone. Step factories
  drop the pattern_fn param. SimpleMLPDecomposedModel implements the method.
- lab: HF_MODEL_FAMILIES registry (model_name -> arch_config/load/model_type)
  in experiments/lm/config.py — the ONLY place a model name selects a family;
  build_target/run_metadata consume it.
- qwen tiny-target tests move to test_qwen3_8b.py (incl. per-layer QK-norm
  pattern test); parity suite re-pointed at the family loader. Full local run
  of affected suites green; make check clean.

Crew-Address: task/ingest-qwen-3-8b
