+++ param_decomp/tests/test_attn_patterns_eval.py
@@ -1,7 +1,7 @@
 """CPU tests for the in-loop attention-pattern recon eval metrics.
 
-Pins the metric-local `attn_pattern_for` target dispatch (shape + causal/softmax sanity on
-both LM targets), the all-false-routes clean target (KL=0 when masked==clean), and the
+Pins the target-owned `attn_pattern` recipe (shape + causal/softmax sanity on both LM
+targets), the all-false-routes clean target (KL=0 when masked==clean), and the
 host-side token-weighted accumulation (combined = Σ sum_kl / Σ n_distributions).
 """
 
@@ -14,7 +14,6 @@ import pytest
 
 from param_decomp.attn_patterns_eval import (
     accumulate_attn_patterns,
-    attn_pattern_for,
     attn_patterns_log_entries,
     make_ci_attn_patterns_step,
     make_stochastic_attn_patterns_step,
@@ -27,9 +26,7 @@ from param_decomp.ci_fn import (
 )
 from param_decomp.components import SiteC, SiteSpec, init_decomp_vu
 from param_decomp.lm import DecomposedModel, run_stochastic_masked_output
-from param_decomp.targets.llama8b import (
-    llama_site_specs,
-)
+from param_decomp.targets.glu_transformer import glu_site_specs
 from param_decomp.targets.llama_simple_mlp import (
     canonical_site_cs as simple_canonical,
 )
@@ -66,16 +63,15 @@ def _build_ci_fn(lm: DecomposedModel, n_embd: int, key: jax.Array) -> CIFn:
     return build_ci_fn(arch, lm.sites, key)
 
 
-def test_attn_pattern_for_shape_and_causal_softmax_llama():
+def test_attn_pattern_shape_and_causal_softmax_llama():
     cfg = _llama_cfg()
-    sites = llama_site_specs(cfg, (SiteC("layers.0.self_attn.q_proj", 4),))
+    sites = glu_site_specs(cfg, (SiteC("layers.0.self_attn.q_proj", 4),))
     lm = _llama_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
-    pattern_fn = attn_pattern_for(lm)
     b, t = 2, 9
     qd, kvd = cfg.n_head * cfg.head_dim, cfg.n_kv_head * cfg.head_dim
     q = jax.random.normal(jax.random.PRNGKey(1), (b, t, qd))
     k = jax.random.normal(jax.random.PRNGKey(2), (b, t, kvd))
-    pattern = np.asarray(pattern_fn("layers.0.self_attn.q_proj", q, k))
+    pattern = np.asarray(lm.attn_pattern("layers.0.self_attn.q_proj", q, k))
 
     assert pattern.shape == (b, cfg.n_head, t, t)
     np.testing.assert_allclose(pattern.sum(-1), 1.0, rtol=1e-5, atol=1e-5)
@@ -84,26 +80,20 @@ def test_attn_pattern_for_shape_and_causal_softmax_llama():
     assert pattern.dtype == np.float32
 
 
-def test_attn_pattern_for_shape_and_causal_softmax_simple_mlp():
+def test_attn_pattern_shape_and_causal_softmax_simple_mlp():
     cfg = _simple_cfg()
     sites = simple_site_specs(cfg, (SiteC("h.0.attn.q_proj", 4),))
     lm = _simple_decomposed_model(cfg, sites, jax.random.PRNGKey(0))
-    pattern_fn = attn_pattern_for(lm)
     b, t = 2, 7
     qd, kvd = cfg.n_head * cfg.head_dim, cfg.n_kv_head * cfg.head_dim
     q = jax.random.normal(jax.random.PRNGKey(1), (b, t, qd))
     k = jax.random.normal(jax.random.PRNGKey(2), (b, t, kvd))
-    pattern = np.asarray(pattern_fn("h.0.attn.q_proj", q, k))
+    pattern = np.asarray(lm.attn_pattern("h.0.attn.q_proj", q, k))
 
     assert pattern.shape == (b, cfg.n_head, t, t)
     np.testing.assert_allclose(pattern.sum(-1), 1.0, rtol=1e-5, atol=1e-5)
     upper = np.triu(np.ones((t, t), bool), k=1)
     assert np.allclose(pattern[:, :, upper], 0.0)
-
-
-def test_attn_pattern_for_refuses_non_attention_target():
-    with pytest.raises(AssertionError, match="only applies to attention targets"):
-        attn_pattern_for(object())
 
 
 def _llama_attn_setup():
@@ -114,9 +104,9 @@ def _llama_attn_setup():
         SiteC("layers.5.self_attn.q_proj", 8),
         SiteC("layers.5.self_attn.k_proj", 6),
     )
-    from param_decomp.targets.llama8b import canonical_site_cs
+    from param_decomp.targets.glu_transformer import canonical_site_cs
 
-    sites = llama_site_specs(cfg, canonical_site_cs(site_cs))
+    sites = glu_site_specs(cfg, canonical_site_cs(site_cs))
     lm = _llama_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
     components = init_decomp_vu(sites, jax.random.PRNGKey(1))
     ci_fn = _build_ci_fn(lm, cfg.n_embd, jax.random.PRNGKey(2))
@@ -125,8 +115,7 @@ def _llama_attn_setup():
 
 def test_ci_step_clean_equals_masked_when_ci_all_one_gives_finite_kl():
     cfg, lm, components, ci_fn = _llama_attn_setup()
-    pattern_fn = attn_pattern_for(lm)
-    step = make_ci_attn_patterns_step(lm, pattern_fn)
+    step = make_ci_attn_patterns_step(lm)
     b, t = 2, 12
     residual = jax.random.randint(jax.random.PRNGKey(4), (b, t), 0, cfg.vocab_size)
 
@@ -142,8 +131,7 @@ def test_ci_step_clean_equals_masked_when_ci_all_one_gives_finite_kl():
 
 def test_accumulate_is_token_weighted_and_combines():
     cfg, lm, components, ci_fn = _llama_attn_setup()
-    pattern_fn = attn_pattern_for(lm)
-    step = make_ci_attn_patterns_step(lm, pattern_fn)
+    step = make_ci_attn_patterns_step(lm)
     res_a = jax.random.randint(jax.random.PRNGKey(4), (2, 10), 0, cfg.vocab_size)
     res_b = jax.random.randint(jax.random.PRNGKey(5), (2, 10), 0, cfg.vocab_size)
 
@@ -173,9 +161,8 @@ def test_accumulate_is_token_weighted_and_combines():
 
 def test_stochastic_step_runs_and_scales_n_by_draws():
     cfg, lm, components, ci_fn = _llama_attn_setup()
-    pattern_fn = attn_pattern_for(lm)
     n_draws = 3
-    step = make_stochastic_attn_patterns_step(lm, pattern_fn, n_draws)
+    step = make_stochastic_attn_patterns_step(lm, n_draws)
     b, t = 2, 8
     residual = jax.random.randint(jax.random.PRNGKey(4), (b, t), 0, cfg.vocab_size)
 
@@ -192,7 +179,7 @@ def test_simple_mlp_step_runs_end_to_end():
     lm = _simple_decomposed_model(cfg, sites, jax.random.PRNGKey(0))
     components = init_decomp_vu(sites, jax.random.PRNGKey(1))
     ci_fn = _build_ci_fn(lm, cfg.n_embd, jax.random.PRNGKey(2))
-    step = make_ci_attn_patterns_step(lm, attn_pattern_for(lm))
+    step = make_ci_attn_patterns_step(lm)
     b, t = 2, 10
     residual = jax.random.randint(jax.random.PRNGKey(4), (b, t), 0, cfg.vocab_size)
 
@@ -293,8 +280,7 @@ def test_attn_patterns_steps_reject_positionless_target():
         leading_axes=(),
     )
     assert lm.leading_axes == ()
-    dummy_pattern_fn = lambda site, q, k: q  # noqa: E731 — never reached; assert fires first
     with pytest.raises(AssertionError, match="LM-only"):
-        make_ci_attn_patterns_step(lm, dummy_pattern_fn)
+        make_ci_attn_patterns_step(lm)
     with pytest.raises(AssertionError, match="LM-only"):
-        make_stochastic_attn_patterns_step(lm, dummy_pattern_fn, 1)
+        make_stochastic_attn_patterns_step(lm, 1)
