+++ param_decomp/tests/test_checkpoint.py
@@ -41,11 +41,11 @@ from param_decomp.lm import DecomposedModel
 from param_decomp.recon import build_loss_terms
 from param_decomp.schedule import ScheduleConfig
 from param_decomp.sharding import hsdp_mesh
-from param_decomp.targets.llama8b import (
-    llama_site_specs,
+from param_decomp.targets.glu_transformer import (
+    glu_site_specs,
     mlp_family_site_cs,
 )
-from param_decomp.targets.llama8b_sharding import (
+from param_decomp.targets.glu_transformer_sharding import (
     init_ci_fn_placed,
     init_decomp_vu_placed,
     init_sources_sharded,
@@ -97,7 +97,7 @@ def _chunkwise_arch(lm: DecomposedModel, cfg: LlamaConfig) -> ChunkwiseTransform
 def _build(seed: int):
     cfg = _tiny_cfg()
     C, seq = 8, 16
-    sites = llama_site_specs(cfg, mlp_family_site_cs(3, 4, C))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(3, 4, C))
     lm = _tiny_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
     vu = init_decomp_vu(sites, jax.random.PRNGKey(seed))
     ci_fn = build_ci_fn(_chunkwise_arch(lm, cfg), lm.sites, jax.random.PRNGKey(seed + 1))
@@ -238,7 +238,7 @@ def _build_sharded(seed: int, mesh: Mesh):
     cfg = _tiny_cfg()
     n = mesh.devices.size
     C, seq = 8 * n, 16
-    sites = llama_site_specs(cfg, mlp_family_site_cs(3, 4, C))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(3, 4, C))
     lm = _tiny_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
     vu = init_decomp_vu_placed(sites, jax.random.PRNGKey(seed), mesh)
     ci_fn = init_ci_fn_placed(
