+++ param_decomp/tests/test_eval.py
@@ -22,7 +22,7 @@ from param_decomp.ci_fn import (
 from param_decomp.components import SiteSpec
 from param_decomp.eval import make_eval_step, next_token_cross_entropy
 from param_decomp.lm import DecomposedModel, run_stochastic_masked_output
-from param_decomp.targets.llama8b import llama_site_specs, mlp_family_site_cs
+from param_decomp.targets.glu_transformer import glu_site_specs, mlp_family_site_cs
 from param_decomp.tests.test_llama8b import (
     _tiny_cfg,
     _tiny_decomposed_lm,
@@ -148,7 +148,7 @@ def test_next_token_cross_entropy_matches_manual():
 def test_eval_step_keys_identities_and_determinism():
     cfg = _tiny_cfg()
     C = 8
-    sites = llama_site_specs(cfg, mlp_family_site_cs(4, 5, C))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(4, 5, C))
     lm = _tiny_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
 
     from param_decomp.components import init_decomp_vu
@@ -219,7 +219,7 @@ def test_eval_step_fresh_pgd_probe():
     random source it starts from (ascent on a fixed objective), and be deterministic."""
     cfg = _tiny_cfg()
     C = 8
-    sites = llama_site_specs(cfg, mlp_family_site_cs(4, 4, C))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(4, 4, C))
     lm = _tiny_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
 
     from param_decomp.components import init_decomp_vu
@@ -281,7 +281,7 @@ def test_eval_step_fresh_pgd_probe_device_count_invariant():
     n_dev = mesh.devices.size
 
     cfg = _tiny_cfg()
-    sites = llama_site_specs(cfg, mlp_family_site_cs(4, 4, 8))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(4, 4, 8))
     lm = _tiny_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
     vu = init_decomp_vu(sites, jax.random.PRNGKey(1))
     ci_fn = _build_ci_fn(lm, cfg.n_embd, jax.random.PRNGKey(2))
@@ -317,7 +317,7 @@ def test_eval_step_l0_groups_sum_member_sites():
     """torch CI_L0 `groups` parity: a group's L0 is the SUM of its fnmatch-member
     sites' L0s; an unmatched pattern refuses at build time."""
     cfg = _tiny_cfg()
-    sites = llama_site_specs(cfg, mlp_family_site_cs(4, 5, 8))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(4, 5, 8))
     lm = _tiny_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
     from param_decomp.components import init_decomp_vu
 
@@ -362,7 +362,7 @@ def test_eval_step_n_valid_rows_masks_pad_tail():
     objective). The stochastic variants draw shape-dependent randomness, so they only agree
     in expectation and are excluded."""
     cfg = _tiny_cfg()
-    sites = llama_site_specs(cfg, mlp_family_site_cs(4, 5, 8))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(4, 5, 8))
     lm = _tiny_decomposed_lm(cfg, sites, jax.random.PRNGKey(0))
 
     from param_decomp.components import init_decomp_vu
