+++ param_decomp/tests/test_source_grad_mean.py
@@ -42,8 +42,8 @@ from param_decomp.ci_fn import Chunk, ChunkwiseTransformerCIArch, build_ci_fn
 from param_decomp.components import init_decomp_vu
 from param_decomp.losses import kl_per_position
 from param_decomp.sharding import hsdp_mesh, shard_batch
-from param_decomp.targets.llama8b import (
-    llama_site_specs,
+from param_decomp.targets.glu_transformer import (
+    glu_site_specs,
     mlp_family_site_cs,
 )
 from param_decomp.tests.test_llama8b import _tiny_cfg, _tiny_decomposed_lm
@@ -56,7 +56,7 @@ def _source_grad(sharded: bool) -> dict[str, jax.Array]:
     reduction we are pinning. Returns one fp32 grad array per site."""
     cfg = _tiny_cfg()
     C, seq, gbatch = 8, 16, 8
-    sites = llama_site_specs(cfg, mlp_family_site_cs(3, 6, C))
+    sites = glu_site_specs(cfg, mlp_family_site_cs(3, 6, C))
     lm = _tiny_decomposed_lm(cfg, sites, random.PRNGKey(0))
     vu = init_decomp_vu(sites, random.PRNGKey(1))
     first_block = min(int(name.split(".")[1]) for name in lm.site_names)
